requirements.txt
pyyaml==6.0.2
requests==2.32.3
python-owasp-zap-v2.4==0.0.14
html2text==2024.2.26
beautifulsoup4==4.12.3